requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
DiscordUtils==1.3.4
dnspython==2.1.0
idna==2.10
multidict==5.1.0
pygicord==1.0.0
pymongo==3.11.4
python-dotenv==0.18.0
pytz==2021.1
requestes==0.0.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
yarl==1.6.3
